auth ---- - [ENH]: Reworked Makefile by defining a variable function that checks the tool version and excluding steps to install tools if not installed. - [DOCS]: Added more content to the READMEs like repo structure and more information regarding different topics like JWT. - [REF]: Removed test credentials for Postgres and extracted them to an .env file and included description for the user in the README. - [UP]: Updated dependencies and Go version to 1.24.4. Also updated Postgres to 17.4.
networking
----------
- [ENH]: Reworked Makefile by defining a variable function that checks the tool version and excluding steps to install tools if not installed.
- [DOCS]: Improved READMEs by giving a description on how gRPC works, especially unary and server streaming gRPCs.
- [UP]: Updated dependencies and Go version to 1.24.4.
- [REF]: Updated implementation of server gRPC streaming, e.g. by removing unnecessary conditions.
