+++ .gitignore
@@ -1,3 +1,13 @@
+# .env
+*.env
+.env.*
+
+# swagger
+swagger.yml
+
+# Go related
+*.pb.go
+
 # Certificate related
 *.pem
 *.csr
