+++ README.md
@@ -1,1 +1,2 @@
 # RoadToDistributedSystems
+RoadToDistributedSystems is my learning journey into the world of backend and distributed systems — exploring design patterns, architectures, and real-world engineering challenges through code.
