--- auth/authentication/.env
@@ -1,5 +1,0 @@
-DB_HOST=127.0.0.1
-DB_PORT=5433
-DB_NAME=auth
-DB_USER=raphael
-DB_PW=test
