+++ auth/postgres-compose.yaml
@@ -1,14 +1,12 @@
 ---
-version: '3.1'
-
 services:
   postgres-auth:
-    image: postgres:15.3
+    image: postgres:17.4
     restart: always
     environment:
-      POSTGRES_USER: raphael
-      POSTGRES_PASSWORD: test
+      - POSTGRES_USER
+      - POSTGRES_PASSWORD
     volumes:
       - ./docker-entrypoint-initdb.d/init.sql:/docker-entrypoint-initdb.d/init.sql
     ports:
-      - "5433:5432"
+      - "${POSTGRES_PORT:-5433}:5432"
