+++ networking/gRPC/server_streaming_rpc/buf.gen.yaml
@@ -1,0 +1,9 @@
+---
+version: v1
+plugins:
+  - plugin: go
+    out: gen/itemstore
+    opt: paths=source_relative
+  - plugin: go-grpc
+    out: gen/itemstore
+    opt: paths=source_relative
