+++ networking/gRPC/server_streaming_rpc/server/go.mod
@@ -1,10 +1,10 @@
-module github.com/RaphSku/RoadToDistributedSystems/tree/main/networking/gRPC/unary_rpc/server
+module github.com/RaphSku/RoadToDistributedSystems/tree/main/networking/gRPC/server_streaming_rpc/server
 
 go 1.24.4
 
 require (
 	github.com/hashicorp/go-hclog v1.5.0
-	google.golang.org/grpc v1.73.0
+	google.golang.org/grpc v1.75.1
 	google.golang.org/protobuf v1.36.6
 )
 
@@ -12,8 +12,8 @@ require (
 	github.com/fatih/color v1.15.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.18 // indirect
-	golang.org/x/net v0.38.0 // indirect
-	golang.org/x/sys v0.31.0 // indirect
-	golang.org/x/text v0.23.0 // indirect
+	golang.org/x/net v0.41.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
+	golang.org/x/text v0.26.0 // indirect
 	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1 // indirect
 )
