--- networking/gRPC/stream_rpc/client/go.mod
@@ -1,20 +1,0 @@
-module github.com/RaphSku/RoadToDistributedSystems/tree/main/networking/gRPC/unary_rpc/client
-
-go 1.20
-
-require (
-	github.com/hashicorp/go-hclog v1.5.0
-	google.golang.org/grpc v1.55.0
-	google.golang.org/protobuf v1.30.0
-)
-
-require (
-	github.com/fatih/color v1.15.0 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.18 // indirect
-	golang.org/x/net v0.10.0 // indirect
-	golang.org/x/sys v0.8.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
-	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1 // indirect
-)
