+++ networking/gRPC/unary_rpc/proto/itemstore.proto
@@ -13,7 +13,6 @@ enum StorageCenter {
 service ItemStoreService {
     rpc CheckStorageAvailability(ItemStoreAvailabilityRequest) returns (ItemStoreAvailabilityResponse);
     rpc DeliverItemToStorage(ItemStoreRequest) returns (ItemStoreResponse);
-    rpc ListStorageLayoutOfCenter(StorageCenterLayoutRequest) returns (stream ItemStoreAvailabilityResponse);
 }
 
 message ItemStoreRequest {
@@ -39,8 +38,4 @@ message ItemStoreAvailabilityResponse {
     int32 availability = 1;
     int32 storage_layer = 2;
     int32 storage_block = 3;
-}
-
-message StorageCenterLayoutRequest {
-    StorageCenter storage_center = 1;
 }
