+++ networking/tls/client/go.mod
@@ -1,10 +1,11 @@
 module github.com/RaphSku/RoadToDistributedSystems/tree/main/networking/tls/client
 
-go 1.20
+go 1.24.4
+
+require github.com/hashicorp/go-hclog v1.5.0
 
 require (
 	github.com/fatih/color v1.13.0 // indirect
-	github.com/hashicorp/go-hclog v1.5.0 // indirect
 	github.com/mattn/go-colorable v0.1.12 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
 	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
